+++ cypress/pages/death/DeathResultsPage.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const ResultsPage = require('./ResultsPage');
+const ResultsPage = require('../ResultsPage');
 
 class DeathResultsPage extends ResultsPage {
